+++ src/index.ts
@@ -2,7 +2,14 @@ import ConfigContext, { IConfig } from './ConfigContext';
 import Config from './Config';
 import AuthContext from './AuthContext';
 import OrdersContext from './OrdersContext';
-import AuthWrapper, { ShowWhenAuthenticated, ShowWhenNotAuthenticated, ShowWhenAuthenticating, ShowWhenNotAuthenticatedAnResolved } from './AuthWrapper';
+import AuthWrapper, {
+  ShowWhenAuthenticated,
+  ShowWhenNotAuthenticated,
+  ShowWhenAuthenticating,
+  ShowWhenNotAuthenticatedAnResolved,
+  IAuthProps,
+  IDeviceLoginState,
+} from './AuthWrapper';
 import OrderProvider from './OrderProvider';
 import SignIn from './SignIn';
 import SignOut from './SignOut';
@@ -20,5 +27,7 @@ export {
   ShowWhenAuthenticated,
   ShowWhenNotAuthenticated,
   ShowWhenNotAuthenticatedAnResolved,
-  ShowWhenAuthenticating
+  ShowWhenAuthenticating,
+  IAuthProps,
+  IDeviceLoginState,
 };